Fix for 3562458, rowspan and colspan converted to ints so that large values will work for them.  Remove 	the ridiculous 1024 limit on the span values.
        * khtml/html/html_tableimpl.cpp:
        (HTMLTableCellElementImpl::parseHTMLAttribute):
        * khtml/rendering/render_table.cpp:
        (RenderTableCell::collapsedBottomBorder):
        * khtml/rendering/render_table.h:
        (khtml::RenderTableCell::colSpan):
        (khtml::RenderTableCell::setColSpan):
        (khtml::RenderTableCell::rowSpan):
        (khtml::RenderTableCell::setRowSpan):
        (khtml::RenderTableCol::span):
        (khtml::RenderTableCol::setSpan):


Canonical link: https://commits.webkit.org/7200@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@8217 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/khtml/rendering/render_table.h
@@ -322,11 +322,11 @@ public:
     long cellIndex() const { return 0; }
     void setCellIndex( long ) { }
 
-    unsigned short colSpan() const { return cSpan; }
-    void setColSpan( unsigned short c ) { cSpan = c; }
+    int colSpan() const { return cSpan; }
+    void setColSpan(int c) { cSpan = c; }
 
-    unsigned short rowSpan() const { return rSpan; }
-    void setRowSpan( unsigned short r ) { rSpan = r; }
+    int rowSpan() const { return rSpan; }
+    void setRowSpan(int r) { rSpan = r; }
 
     int col() const { return _col; }
     void setCol(int col) { _col = col; }
@@ -386,8 +386,8 @@ protected:
     
     int _row;
     int _col;
-    ushort rSpan;
-    ushort cSpan;
+    int rSpan;
+    int cSpan;
     int _topExtra : 31;
     bool nWrap : 1;
     int _bottomExtra : 31;
@@ -406,8 +406,8 @@ public:
 
     virtual const char *renderName() const { return "RenderTableCol"; }
 
-    long span() const { return _span; }
-    void setSpan( long s ) { _span = s; }
+    int span() const { return _span; }
+    void setSpan(int s) { _span = s; }
 
     virtual void addChild(RenderObject *child, RenderObject *beforeChild = 0);
 
@@ -424,7 +424,7 @@ public:
 #endif
 
 protected:
-    short _span;
+    int _span;
 };
 
 };
